modified:   ../Caesar-cipher-attack/FreqAnal.h 	modified:   ../Caesar-cipher-attack/main.cpp 	modified:   Caesar-Shift.h 	modified:   main.cpp

+++ Substitution-Ciphers/Caesar-Shift.h
@@ -10,7 +10,7 @@ public:
 
     void setPlainText(const char* text) {
         int i=0;
-        for(; i < strlen(text); i++)
+        for(; i < (int)strlen(text); i++)
             plaintext[i] = toupper(text[i]);
         charCount=i;
     }
